more tests + code cleanup
added more thorough tests, some of which are done at compile time for validating return types and intrinsics.
some of the tests are for acting on const mutex, which was supported in previous commit but was not tested.

merged the guard and mut_guard code into one templated type that we typedef out.
got rid of the silly rvalue deleted functions - could technically be safe even on prvalue if all done in same top level expression.
moved the internal logic of locking into the guard object, which now just takes the mutex by reference.